+++ apps/pleiepenger-sykt-barn/e2e/cypress/integration-utils/contextConfig.ts
@@ -48,5 +48,6 @@ export const contextConfig = (props?: ConfigProps) => {
             'postInnsending',
         );
         cy.intercept('*.api.sanity.io', {});
+        cy.intercept('*amplitude*', {});
     });
 };