+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @navikt/omsorgspengerutbetaling-arbeidstaker-soknad
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [86d8f70]
+  - @navikt/sif-common-amplitude@2.24.15
+
 ## 1.0.3
 
 ### Patch Changes