+++ apps/pleiepenger-i-livets-sluttfase-soknad/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @navikt/pleiepenger-i-livets-sluttfase-soknad
 
+## 2.5.22
+
+### Patch Changes
+
+- Updated dependencies [86d8f70]
+  - @navikt/sif-common-amplitude@2.24.15
+
 ## 2.5.21
 
 ### Patch Changes